+++ spec/spec_helper.rb
@@ -191,6 +191,7 @@ RSpec.configure do |config|
     Deimos.config.reset!
     Deimos.configure do |deimos_config|
       deimos_config.producers.backend = :test
+      deimos_config.schema.nest_child_schemas = true
       deimos_config.phobos_config_file = File.join(File.dirname(__FILE__), 'phobos.yml')
       deimos_config.schema.path = File.join(File.expand_path(__dir__), 'schemas')
       deimos_config.consumers.reraise_errors = true
@@ -199,7 +200,7 @@ RSpec.configure do |config|
       deimos_config.logger = Logger.new('/dev/null')
       deimos_config.logger.level = Logger::INFO
       deimos_config.schema.backend = :avro_validation
-      deimos_config.schema.generated_class_path = 'spec/schema_classes'
+      deimos_config.schema.generated_class_path = 'spec/schemas'
     end
   end